feat(api-service): layouts content variable and preview rendering
