+++ apps/api/src/app/layouts-v2/usecases/layout-variables-schema/layout-variables-schema.usecase.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InstrumentUsecase } from '@novu/application-generic';
 import { JsonSchemaTypeEnum } from '@novu/dal';
+import { LAYOUT_CONTENT_VARIABLE } from '@novu/shared';
 
 import { JSONSchemaDto } from '../../../shared/dtos/json-schema.dto';
 import { buildSubscriberSchema } from '../../../shared/utils/create-schema';
@@ -15,6 +16,9 @@ export class LayoutVariablesSchemaUseCase {
       properties: {
         // TODO: implement subscriber data schema
         subscriber: buildSubscriberSchema({}),
+        [LAYOUT_CONTENT_VARIABLE]: {
+          type: JsonSchemaTypeEnum.STRING,
+        },
       },
       additionalProperties: false,
     };
