+++ packages/shared/src/consts/index.ts
@@ -11,3 +11,4 @@ export * from './inviteTeamMemberNudge';
 export * from './preferences';
 export * from './validIdRegex';
 export * from './feature-tiers-constants';
+export * from './layouts';
