+++ packages/shared/src/consts/layouts.ts
@@ -1,0 +1,1 @@
+export const LAYOUT_CONTENT_VARIABLE = 'content';
